[home_work_2] add logic show movies details from server

+++ app/src/main/java/ru/androidschool/intensiv/data/moveid/MovieId.kt
@@ -1,4 +1,4 @@
 package ru.androidschool.intensiv.data.moveid
 
 @JvmInline
-value class MovieId (val id: Long)
+value class MovieId (val id: Int)
